Excluding auto-generated classes from test coverage
Since we cannot properly test the generated classes (e.g. the annotations seem to be always uncovered), we exclude these generated sources. This is usually also best-practice.
